Some cleanup related to QuarkusConsole
- Use SmallRye Common to determine the OS rather than custom code
- Avoid initializing QuarkusConsole when setting up the logging layer,
  as all we need in prod mode is to determine the color status,
  QuarkusConsole has some additional static fields that are only useful
  in dev mode

Franky, I don't think we should use dev mode SPI code in prod code, but
I'm not sure where we could move this terminal detection API, maybe to
something in SmallRye Common?

(cherry picked from commit 9c615445d091606079560eb95eae7cb61886c117)
